pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-cooperation"
version = "1.0.0"
description = "Intelligent multi-model LLM cooperation system with OpenAI API support"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "LLM Cooperation Team", email = "team@llmcooperation.com"}
]
maintainers = [
    {name = "LLM Cooperation Team", email = "team@llmcooperation.com"}
]
keywords = [
    "llm", "cooperation", "ai", "openai", "api", "routing", 
    "multi-model", "inference", "nlp", "chatbot", "assistant"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research", 
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9", 
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "openai>=1.12.0",
    "aiohttp>=3.9.0",
    "asyncio-mqtt>=0.13.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "pyyaml>=6.0",
    "click>=8.1.0",
    "rich>=13.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.12.0", 
    "isort>=5.13.0",
    "flake8>=6.1.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0"
]
server = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "prometheus-client>=0.19.0"
]
monitoring = [
    "psutil>=5.9.0",
    "GPUtil>=1.4.0", 
    "structlog>=23.2.0"
]
all = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "prometheus-client>=0.19.0",
    "psutil>=5.9.0",
    "GPUtil>=1.4.0",
    "structlog>=23.2.0"
]

[project.urls]
Homepage = "https://github.com/llm-cooperation/llm-cooperation"
Documentation = "https://llm-cooperation.readthedocs.io/"
Repository = "https://github.com/llm-cooperation/llm-cooperation.git"
"Bug Reports" = "https://github.com/llm-cooperation/llm-cooperation/issues"

[project.scripts]
llm-cooperation = "llm_cooperation.cli:main"
llm-config = "llm_cooperation.tools.config_cli:main"
llm-server = "llm_cooperation.server.main:main"

[tool.setuptools]
packages = ["llm_cooperation"]

[tool.setuptools.package-data]
llm_cooperation = [
    "config/*.json",
    "templates/*.yaml", 
    "static/*"
]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["llm_cooperation"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers"
testpaths = [
    "tests",
]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
source = ["llm_cooperation"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "setup.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]